Extract class Api::Settings
While the api.yml file is parsed at the class level, most of the work
that interacts with it is done at the instance level of the
ApiController. This means that any object other than the ApiController
that wants to query the API config will have to duplicate this
code. This is a first step towards untangling it.


(transferred from ManageIQ/manageiq@961141150c289bd82e60084eb674f7319c66eef5)