Add test for fuzzy matching on strings with /'s
This revision originally started life as a workaround, which lead to the
discovery of CVE-2016-7040. Since
https://github.com/ManageIQ/manageiq/pull/12822, which addressed the
security issues, has obviated the need for a workaround on the API side,
this revision serves only to prove that this feature of the API works
properly, preventing future regressions.

Closes https://github.com/ManageIQ/manageiq/issues/11035


(transferred from ManageIQ/manageiq@82eabfd8f54cd1da2423d37a311780a809efa923)