Move ApiController under Api
Initially it was desired that

```
GET /api
```

would route to `ApiController#index`, and

```
GET /api/vms
```

would route to `Api::VmsController#index` (or `#show`, as it is
currently).

Unfortunately this doesn't provide an easeful solution for
configuration, so it seems to make sense to pull it in under the `Api`
namespace - the main benefit being that, should it need to access
constants under the `Api` namespace they won't need to be fully
qualified (so it can look and behave just like all the other API controllers).


(transferred from ManageIQ/manageiq@fe69cc433ec75886a83af3e4bb6b8af62f5a2a5b)