Provide an API to set headers ahead of request
The current implementation made setting the "HTTP_AUTHORIZATION" header
declaratively quite awkward - it involved setting an instance variable
and merging in the header at request time (i.e. with `run_get`, etc.),
depending on whether that variable was set.

The alternative proposed in this revision is to provide a memoized
`headers` variable which contains the default headers used in every API
request, allowing for the authorization header to be set where it is
created in the declarative `basic_authorize`.


(transferred from ManageIQ/manageiq@12bfe2b4cbb7aa98aa987a1bf66980cea0167f35)