Use proper MiqExp operators for datetime filtering
Filtering results in the API on date/datetime attributes can be done,
but there's not exactly a 1-1 relationship between the syntax of the
filtering API and that of MiqExpression, which supports it.

Using comparison operators to filter on dates just *happens* to work
because arel, which is in turn supporting MiqExpression, knows how to
deal with it. But it's not officially supported - the documented way to
form expressions with datetimes is by using the date time operators:
"IS", "BEFORE", "AFTER", "FROM", "IS EMPTY", and "IS NOT EMPTY". Using
an unsupported operator in an expression that has to be evaluated in the
context of Ruby (i.e. it can't be completely evaluated in SQL because of
the presence of one or more virtual attributes) will break.

The best we can do most immediately is to provide a simple mapping for a
subset of the operators:

| Filter API | MiqExpression |
| ---        | ---           |
| `>`        | `AFTER`       |
| `<`        | `BEFORE`      |
| `=`        | `IS`          |

Anything else will result in a bad request, which is better than a 500.

Fixes https://github.com/ManageIQ/manageiq/issues/12648


(transferred from ManageIQ/manageiq@75a26368eb8ea8cbd2bb5cbe76c9f4e8ef91aa6e)