Respond 400 if filtering is unsupported
With certain subcollections (e.g. vms) the object returned is an array,
not a relation. That means we can't tack on additional scopes as we do
when filtering. Since we can't easily enforce returning relations for
all subcollections, the best we can do is to return a 400 instead of a
500, indicating that filtering is not supported for affected
subcollections.


(transferred from ManageIQ/manageiq@063df2a6ae9f490fd327d6d94f55646bf1159161)