Add alert statuses api
GET /api/alerts
GET /api/alerts/:id

Adds an api to view MiqAlertStatuses The entity is renamed to `alerts`
since that is thought to be the most useful name for users. When adding
MiqAlert in the future the intention is to use the name
`alert_definitions` - that is consistent with
miq_event/miq_event_definition.


(transferred from ManageIQ/manageiq@0476bc3a6ff0953b736e413ea3856664eebcd1c5)