Re-add logging the action, controller, format
In #13257 logging was changed in a way which omitted the action,
controller and format from being logged for the sake of restoring the
build to green.  This revision restores that behavior and adds testing
to ensure that this data gets logged.


(transferred from ManageIQ/manageiq@3e5441d59d1bf145bea0f9f1fcd28777f88a7f0c)