Inherit from ActionController::API
`ActionController::API` is a more stripped down version of
`ActionController::Base` designed just for APIs (as the name
suggests). We use basic authentication in our API, so it's necessary to
include the
`ActionController::HttpAuthentication::Basic::ControllerMethods` by
hand. On the plus side, this gives us everything we need and no more.

The initial motivation for this change was to fix the Brakeman warning
that resulted from un-inheriting `ApplicationController`. Brakeman uses
static analysis to try to look for possible security issues, and found
one in `protect_from_forgery` not being declared anywhere. (It wasn't
smart enough to discover that the hook that this registers was being
unregistered in the API controllers, so its important to note that
nothing has changed in terms of outward behavior with regards to
`protect_from_forgery`). The way Brakeman has dealt with this issue is
to only complain if your controller inherits from
`ActionController::Base` (which `ActionController::API` doesn't).

See: https://github.com/presidentbeef/brakeman/pull/611


(transferred from ManageIQ/manageiq@16a1aa48dc3c08ea2c23991eff3c3fcd83c66390)