Preserve contract for expressions of alert definitions
This requires setting the `@additional_attributes` to include
`"expression"` for GET and POST (update) requests. For POST (create)
requests, because the object is wrapped in a "results" object, we need
to do a bit of a hack to insert the expression into a serialized
version of the alert definition before it gets rendered.