Express ids as uncompressed strings
Support for compressed ids was added and enabled by default (rendering
all ids in their compressed form) with the thought of killing two
birds with one stone:

1. javascript doesn't support very large numbers
2. we wanted to add support for compressed ids

It turns out that (2) wasn't solving any problems in UI land
anymore. With the added complexity and emerging issues¹ that came as a
result of this change, it seemed better to remove the compressed id
support and simply render all ids as strings instead, which is
sufficient to solve the javascript large numbers issue.

Support for incoming compressed ids in URLs has been left, since this
has already been incorporated into the Fine release.

1. See https://github.com/ManageIQ/manageiq-api/pull/49,
https://github.com/ManageIQ/manageiq/pull/15658