Avoid extra string allocations in #parse_href
Pedantic commit, but basically calling "string#{}" and `.sub` is
generating a new string instead of making use of the one that is already
in place.  A small attempt to reduce the object allocations for this
method (also, steps toward making this file `frozen_string_literals`
compatible).

This does require doing a `href.dup` in the first line to avoid
modifying the original href value (if that happens to be used
elsewhere), but should overall reduce the new strings created in this
method.