Added support for arbitrary resource identifier.
Generally, resources are accessed by their id attribute.  For
automate_workspaces, this is not the case as we only want
resources accessed by their guid.

The enhancement adds support for arbitrary resource identifier
declared in the collection definition in the api.yml.
hrefs created for the resource themselves as well as the actions
made available to those resources will have the guid (as defined
by the :resource_identifier: paramter) in the href and not the id.

With the href properly built, the clients leveraging them will
continue to work as they do today.