Use HTTP status to indicated that cids have been deprecated
Using the HTTP status is the Right Way :tm: to show that a resource
has been moved (which, to the outside world, it has - it has a
different id). With this change we can remove some of the handling
done in the request parser that checks for compressed ids. We still
need some of the custom handling, however, since we still support
compressed ids in hrefs in the request body, which cannot use the same
strategy.