DRY up version regex
We currently have two things claiming to be the version regex -
`Api::VERSION_REGEX` and `ApiConfig.version.regex`. It seems sensible
to remove one of them. Somewhat alarmingly they are not exactly the
same, but the tests do pass. If it turns out we do need two different
ones, we should probably:

* put them in the same place
* use better naming to convey intent
* add tests that cover the unique behavior