Fix failing tests from changed ovirt provider
The manageiq-providers-ovirt had a new change recently that caused
issues with the tests:

https://github.com/ManageIQ/manageiq-providers-ovirt/commit/87703c6

This causes the provider tests to fail because the ovirt provider does
not properly handle a missing/not present SDK when asking for api
versions by default.  This adds a stub to support that.

Note:  We returning the v3 version from the stub since the :sample_rhevm
attributes are only setup to work with that version of the API, and it
fails to work with the v4 adapter.