Return action responses for bulk delete and not found for single resources
The way deleting works now is that resource search happens prior to the rescue block, which is causing an issue when the first item in bulk deletion happens, it raises a not found when it should just be an action response. This PR leverages the existing delete_action_handler to ensure that not founds get returned for single resource actions, but action results are returned for bulk actions.

https://bugzilla.redhat.com/show_bug.cgi?id=1504693