Disable custom actions on /services
This is a temporary fix for https://bugzilla.redhat.com/show_bug.cgi?id=1499692

Including custom actions is incredibly slow due to the implementation of
custom buttons and how delegations happen, as well as a lack of
preloading on the API side. This causes massive N+1
queries and ultimately makes the query on /services 4-5x slower.

For now, disable custom actions on services only - but the entire
feature should be reviewed in the future.