Unmeta/expand router
This commit completely expands the metaprogramming based on the api.yml
in the routes. It's a complete 1:1 expansion with absolutely no changes
in route order or even differing ActionDispatch calls at all from the original.

Note that this is the beginning and not the end result of using more
standard Rails conventions - things can and should be combined in to
`resource` style calls - but to get started and avoid any future
conflicts this is a good start.