Use destroy_queue for provider delete.
Changes to ExtManagementSystem now require the use of destroy_queue to queue the deletion of the provider. Without this change, provider deletion is silently failing due to a before_hook on destroy that checks if there are any active queues for the provider. The orchestrate_delete method that will now be called will resolve this.

https://bugzilla.redhat.com/show_bug.cgi?id=1501941