Add subresource_action_identifier spec helper
The current subcollection_action_identifier does not account for subresource actions, which meant that the subresource action could not have a different identifier than the subcollection action. This adds in a new subresource_action_identifier which takes into account both subresource actions defined on the type and on the subtype. This will now allow the orchestration template sub collection actions to be identified properly. Also adjusts a snapshots spec to ensure that the helper works for both cases of identifiers.