Return correct subcollection actions
Right now, sub collection actions that are specified on a specific collection (i.e., :generic_object_subcollection_actions: specified on :services: collection) are not taking precedence to those sub collection actions defined on the collection itself (i.e. :subcollection_actions: defined under :generic_objects:). This is causing incorrect actions to be returned, causing the users to believe that they have actions that they do not, such as :create: for the generic objects sub collection on services.

https://bugzilla.redhat.com/show_bug.cgi?id=1518833