Ensure that correct target is passed to resource_search on dialog refresh
Currently, the API assumes that a pluralized target_type will match that of a collection class type. This is not the case, as can be seen from “service_template_ansible_tower”. This instead takes the target type, camelizes and constantizes it in order to get the correct name for the class. Resource search is then able to proceed normally. In addition, this will raise a bad request error for invalid target_types rather than raising an internal server error

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=1524227

Update name_for_subclass to work with strings to remove the constantize that is present on user parameters in service_dialogs_controller