Fix custom actions hrefs
Because custom actions are represented as a plain hash in the model,
we lose all context about the type of object we must add an href to in
the normalization process. This process tries, in lieu of this
context, to guess what the href path should look like, and is
incorrectly linking these objects to the parent object's
API (e.g. generic objects, services, etc..). Refactoring all of
normalization to get the proper context will be necessary in order to
fix this correctly. Instead, we intercept the result here, adding the
correct hrefs, which will not be overwritten later.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1518297