Fix collection override
To fix hrefs in base controller abstraction land that doesn't have
concrete controllers/serializers for exceptions, we've 'normalized'
hrefs the best we could by deterministically giving an href based on the
request that prefers to be nested when possible - but are now screwed
because it's actually determined by what you can and can't do with
nested collections in the interface.

So, this does the same sort of hacky override that was here, just
without the misleading helper that was present before by constructing
our own URL deliberately.