Settings CRUD on Servers, Zones, and Regions
This adds manipulation of all settings on servers, zones, and regions
via the API in core's `ConfigurationManagementMixin` in the form of a new
base controller method, `settings`, which routed to these three
collections using the `settings` options keyword in the YAML.

Note also that some overrides had to be added to circumvent the
strictness of our 'validations' on HTTP methods for collections. The
actions against this path (/settings) under a collection do not
correspond to the actions you can take against a collection (the ones
specified in the YAML), so we ignore validating HTTP methods if you're
looking to do settings stuff specifically. This, by the way, is a great
example of 'doing things that aren't collection CRUD is really, really
hard'.

The interface here is very simplistic and just follows the API provided
by the core mixin. The responses (and requests, for that matter) lack
any sort of envelope (as again, going through the normal API rendering
stuff doesn't really make sense), which I'm not sure is good or bad.

GET /servers/:id/settings
Renders the server settings as provided by the settings API.

PATCH /zones/:id/settings, body: {"ems": { "ems_ansible_tower": "blacklisted_event_names": ["EventName1"] } }
Sets the ems.ems_ansible_tower.blacklisted_event_names to ["EventName1"]
This action serves as both create and update

DESTROY /zones/:id/settings, body: ["ems", "ems_ansible_tower"]
This deletes the settings change for this zone under this key. All
changes underneath this key path would be reset and the settings for
this zone's ems.ems_ansible_tower subtree would revert to whatever is
higher up the settings tree for this zone.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1486634