Allow additional provider parameters to be specified in the API
This update will allow providers to have more control over the parameters that are allowed when creating a new manager. Currently, only those columns for a provider are allowed. This will also be a good transition starting point to put more of the validation inside of the providers themselves, rather than store it all within the API controller.