Create Validator
The parser is currently overloaded with validation logic that is difficult to follow. This moves the validation into smaller, more concise methods. It also creates more methods to handle a lot of the request configuration logic while utilizing memoization, as a lot of calls were repetitive and the naming of variables were very difficult to follow.