Update role identifiers for tasks collection to be the same as those used in the UI
When user roles are seeded, they have either the miq_task_all_ui feature or miq_task_my_ui feature (or both), which is what the UI also uses to validate that they can view requests. However, the API uses tasks_view. This is causing a problem where users that are able to see tasks in the UI are unable to see tasks through the API. This updates the product features used in the API to match what is used in the UI.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1535962