Return sui product features for all of a user's groups on the API entrypoint
This is needed so that the users can know what features they will have before changing groups. When they are a part of multiple groups, the current way of getting groups /api/groups may not be available to them based off of their current groups permissions. This will fix that.