Default section to "metadata" across custom attributes add action
When a new custom attribute is added, it defaults the section to "metadata" when the section is not specified. This causes issues if the request is then made again, especially because the default behavior for "add" is to edit. Without defaulting the section to "metadata" across the board, a new custom attribute is created, rather than updated.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1544800