Always challenge a user with no/bad credentials
According to RFC 2617:

> The 401 (Unauthorized) response message is used by an origin server
  to challenge the authorization of a user agent. This response MUST
  include a WWW-Authenticate header field containing at least one
  challenge applicable to the requested resource.

We currently only issue a challenge when no credentials are provided -
if a user fails authentication they are not challenged again. This
broken implementation can cause some frustration for instance when
using the web browser, which will cache a mis-typed user/password
combination and not challenge the user again.

This is somewhat problematic as we only use the `Authorization` header
for Basic Authentication. It seems like a safe compromise for the user
to be challenged to use Basic Authentication if authentication fails
for whatever reason.