Remove product feature needed to access automate workspace
https://bugzilla.redhat.com/show_bug.cgi?id=1598464

Each automate workspace is identified with a unique guid
and it is only available for the lifetime of the method.
When this feature was initially added we use 'miq_ae_domain_view'
product feature, which is only available for admin user and
for non admin users we get a 403 Forbidden error because they dont
have the miq_ae_domain_view feature enabled. This PR does away
with the miq_ae_domain_view feature to be enabled for each user
to access the workspace.