[specs] Fixes conversion_hosts_spec.rb failures
A recent change in https://github.com/ManageIQ/manageiq/pull/18277
caused specs to fail in the API since the new validations no longer
allow the specs to create `:conversion_host` factories.

This change simply updates the specs to support these new restrictions,
but it might not be the ideal API interface.