Filter - add a new ===, !== operators to match exact strings without mangling
Right now, if you have a VM (or anything) with a % or * in name,
it's impossible to find via the API by that name, without using wildcards.

This adds 2 new operators: `===` and `!==`, which do the same thing as `=` and `!=` without trying to interpret the value as a regular expression.

The filtering code still mangles whitespace, but wrapping the value in single quotes prevents that.