Log exceptions to rails console
Right now, if any UI request fails with an exception, it can be seen in rails console output and in development.log / production.log.
But, if the same problem happens with an API request, the only place where that exception is visible is api.log.

(This is the same code as in UI, except for an extra condition to only show the backtrace for internal_server_error class of bugs.)

Logging API exceptions to rails output as well.