Removing the disabled flag in the api.yml configuration file.
This was added during the infancy of the API when we identified
actions we wanted to implemented and obtained the related
entitlement from the product features file, but did not yet
implement them or were never asked for.

While it was useful to have, it does cause confusion while
reading the api.yml and adds some complexity to the code base.
This is no longer needed as the API has expanded since and has been
enhanced over the years and actions added on an as-needed basis.