[BaseController] Allow counts with .select
When including a `virtual_attribute` that can be included via SQL (aka,
`.select`), the `.count` function doesn't work because it doesn't build
the "counting" function properly.

For now, the simple solution is to just modify the scope where any
counting might occur via the `Api::QueryCounts` in the base controller
by simply removing the `.select` value from the `scope`, and allowing
the counting to happen after that.