[Renderer] Allow nested attributes for include_for_find
Modifies the determine_include_for_find method to support nested
relationships, but requires a bit extra to do so by converting the
resulting array to a Hash that ActiveRecord::Base#includes can
understand.

Also, in the specs, OperatingSystem and ComputerSystem were added to
each Vm since they are classes that don't participate in Rbac, but have
a nested relation to Vm.