[Renderer] Handle :uses for :include_for_find
Makes use of the `.virtual_includes` to determine what relationships
need to be preloaded to for `virtual_columns` that make use of the
`:uses` query.

Unfortunately, there are multiple permutations of how value of `:uses`
can be assigned:

- As a `String`/`Symbol` representing a single `ActiveRecord` relation
- An `Array` of relations that are based off the base model.
- A `Hash` that represents a tree of relations starting from the base
  model

Since each of these need to be supported, there needs to be a common
data structure that they all can be converted to in the case that
multiple attributes are included, and `Hash` is that data structure.

Thankfully this is mostly solved via `activerecord-virtual_attributes`
and its `.merge_includes` method, so most of this is pretty straight
forward once that is in place.