[Api::BaseController::Generic] add_subcollection_data_to_resource + edit
Allows `add_subcollection_data_to_resource` to be used with update/edit
actions in a generic fashion.

This already existed previously on create, with subcollections be able
to define `<subcollection>_assign_resource` methods which would get
called if the proper resources were defined.  However, `edit_resource`
did not support this, and so this would be sent directly to `.update!`
and would not work.

This change enables this for `edit_resource` in the same fashion that it
has existed for `add_resource`.