[Subcollection::Authentication] Allow subcollection assignment
Allows assign authentications via a subcollection relation.

For example, a zone can edit it's authentication relation by doing:

    # POST /api/zones/X
    {
      "action": "edit",
      "resource": {
        "authentications": [{
          "href": "/",
          "windows_domain": {
            "userid": "foo",
            "password": "bar"
          }
        }]
      }
    }

Classes that include this Subcollection will also have to implement the
`.valid_authentication_types` so that they can validate the types of
keys allowed.  Of note:  there should be only one key per auth, minus
the `"href"` key, which is removed from the data hash prior to sending
it to the AuthenticationMixin in ManageIQ/manageiq.