[Renderer] Fix duplicate includes with extra_cols
When using just `klass.virtual_includes`, it only checks to see if the
given `virtual_attribute`/`virtual_column` also includes associated with
it.  On the surface, that seems like a sufficient check, but since a
`virtual_attribute` can have a `:uses` associated with it, as well as
`:arel`, that means the column could be included as both an entry to
`:extra_cols`, and for `:include_for_find`.

The problem with certain columns for `:include_for_find` is that they
can very easily lead to "LEFT JOIN bombs", as in the example shown in
the specs:

`virtual_total :total_vms` and `virtual_aggregate :aggregate_vm_cpus`
both use the `:vms` relationship on `ExtManagementSystem`, but the
latter also has a secondary relationship to hardware, and is aliased as
`:vm_hardwares` as a `.includes` target.

As a result, both of those two columns combined in this fashion
previously caused the `vms` table to be `LEFT OUTER JOIN`'d twice,
meaning it included all of the columns of that table (60+) twice for
each row.  On a large dataset, with LEFT OUTER JOINS also causing a huge
matrix of multiple number of duplicate rows meaning the SQL query for
even a system with 1000+ Vms can be crippled with this one query.

This change just ensures that we are avoiding the `:include_for_find`
when that will be present and gathered via a `:extra_cols` elsewhere,
hopefully avoiding some of these "LEFT JOIN bombs" as a result.