[Renderer] Use next, not return
In the context of this block, using `return` will eject from the
iteration of the block, AND also the method, so in all of usages of
`return` previously, it will `nil`, which is a valid return value for
the method.

With `next`, it will just skip anything after the `yield` and move on to
the next item in the iterator (which isn't obvious we are in one with
this code change).  This is the behavior that was intended, but because
the specs around this allow skipped at the proper times with the
previous implementation, it wasn't know that this was a problem.

`next` is also used in the `determine_include_for_find` method, so this
should be a relatively safe and tested change.