[requests/providers_spec.rb] Jansa approved spec
There has been a bit of a saga with these specs:

- https://github.com/ManageIQ/manageiq-api/pull/933#issuecomment-715313395
- https://github.com/ManageIQ/manageiq-api/pull/933#issuecomment-716787919

And it mostly ties in with some code that was updated for kasparov and
up for `aggregate_vm_cpus`, but wasn't backported to `jansa` (which was
probably the right choice).

However, it caused the specs to behave differently on the two branches,
and thus fail on `jansa`.  The nice part about these originally is that
these specs very much replicated what we were seeing here:

https://github.com/ManageIQ/manageiq-api/issues/923#issuecomment-713160644

But since that isn't valid for both branches, it is better to use specs
that are suited for both branches.