[WIP][api_config.yml] Fixes :indentifier: updates
Changes were made in Core:

https://github.com/ManageIQ/manageiq/pull/20438

That updated the product feature `:identifier:` key/values to support
the Control Explorer split.

This caused the `spec/lib/api/api_config_spec.rb` to fail when it
checked the difference between the `api_config.yml` and the configured
`MiqProductFeature` records to confirm they are in sync.

This trues up the config with the new changes from the aforementioned PR
above to allow the specs to pass.