Make api debuggable from browser again, accept cookie only for GET
HTTP Basic is no longer really supported by the API, without the auth request, the browser ignores URLs like http://admin:smartvm@localhost:3000
So, there is currently no way to open an api request directly from the browser window, for debugging.

Since there was presumably a reason to break http basic that way (it should still works if the client ignores it and sends the right headers, so it does work for curl),
but there's also reason to make this accessible from the browser, adjusting the ui_session code to:

* treat any request with the right *cookie* as an UI request, not only when CSRF present
* don't require CSRF token for GET requests

That makes it possible to open an API tab as long as the user is logged in into manageiq UI in the browser.