[ZonesController] Allow "setting" settings
The Zone model/api-controller has two overriding features:

- A subcollection for "settings"
- A hash column called "settings"

The code in `Api::BaseController::Generic#add_resource` has a block in
it that takes any key from the params hash that is part of a
subcollection, and passes it to create those sub-resources.  This
overrides the behavior of being able to write the `settings` hash
directly on create (this is not a behavior that is replicated
generically for `edit_resource`)

To allow for editing of `Zone#settings` on create, we override
`add_resource` so that it will not include this sections of code for
this particular controller.

Note:  The `alias` also needed to be re-applied, as `create_resource` is
what is actually called by `Api::BaseController::Generic#create`