[BaseController::Render] Fix subquery_res.count
Very simply, the API can be provided with a query that can be reduced
down to something like this:

  Vm.select(Vm.arel_table[Arel.star], :provisioned_storage).count

This, in Rails 6, does not seem to work, as a `String#join` is done to
the `Vm.arel_table[Arel.star]` and the `:provisioned_storage` columns
deep within the `ActiveRecord` library, and ultimately the
`Vm.arel_table...` portion will be converted into a `struct` when passed
as a SQL string, and not a column name or function, looking something
like this:

  SELECT COUNT(#<struct Arel::Attributes::Attribute relation=#...

This obviously isn't valid SQL, and while it might make sense to fix
this in Rails (as it seems to be a bug there), the solution isn't
terribly obvious, and this is a simple fix on our end.

By excluding :extra_cols from the from the options when defining
`subquery_res`, we avoid any `.select` calls for being applied in Rbac.
Further more, we can't use `.except(:select)` on the resulting
`subquery_res` object because it can also be an Array, not just an
`ActiveRecord::Relation`, so avoiding `.select` in the first place
simplifies the logic.

* * *

Side note:  The following also doesn't seem to be valid:

  irb> Vm.select(:id, :provisioned_storage).count
  (0.5ms)  SELECT COUNT(vendor, id) FROM "vms" WHERE ...
  PG::UndefinedFunction: ERROR:  function count(character varying, bigint) does not exist
  LINE 1: SELECT COUNT(vendor, id) FROM "vms" WHERE "vms"."type" IN ($...
                 ^
  HINT:  No function matches the given name and argument types. You might need to add explicit type casts.

So it seems like the logic here is possibly flawed is Rails in general,
but what lead to this isn't clear.