[Subcollection::AlertDefinitions] Fix unassign success
The change from using RelationshipMixin for MiqSet to using a dedicated
join table caused the return value for `.remove_member` to change from
returning an `ActiveRecord::AssociationRelation` (`Array` of results) to
using a `.delete_all`, which returns just an `Integer` (which is the
count of the affected records).

To also account for any `ActiveRecord` errors, a `rescue` was added to
give a proper response for those as well.