[users_spec.rb] Fix test order failures
If the `authentication_spec.rb` is run before the `user_spec.rb`, it can
set the `TokenStore.token_caches` to a particular value ("sql" in
particular) that doesn't match what is required for the test (based on
the `session_store` value), and causes the tests to fail.

By clearing the `TokenStore.token_caches` value before an after each
test, we avoid this failure.

The test failures can be reproduced in the previous commit with the
following rspec command:

  $ bundle exec rspec --order defined spec/requests/authentication_spec.rb:253 spec/requests/users_spec.rb:633